[dynamo] Account for function id reuse in relevant Dynamo decorators
This fixes a recent series of flaky failure from `nonstrict_trace` unit
tests: #148166, #148056, #148055, #148054, #148034, #148033, #148032, #148031.

For now we don't need to worry about the other decorators because they
are either meant for builtin/numpy functions (which should never
deallocate in practice), or used for polyfills which keeps the function
object in `get_torch_obj_rule_map()`.

Fixes #147777.

ghstack-source-id: 2edebd4599552b210944524d5ba88a4140d6f38b
Pull Request resolved: https://github.com/pytorch/pytorch/pull/148385